Fix benchmark analyzer violations and refresh results
Analyzer cleanup:

- Remove the added benchmark warning suppression list and the file-level pragma.

- Split ReactiveExtensionsComparisonBenchmarks into documented partial files under analyzer line-length limits.

- Replace the compiled regex with GeneratedRegex and fix task, ternary, naming, ordering, magic-number, and unused-member diagnostics directly.

Benchmark refresh:

- Rerun the joined 610-benchmark BenchmarkDotNet suite after the benchmark harness changes.

- Update README benchmark artifact paths, duration, comparison posture, and all 87 ReactiveUI.Primitives.Extensions rows.

Validation:

- dotnet build src/ReactiveUI.Primitives.slnx -c Release passed with 0 warnings and 0 errors.

- Extension TUnit executables passed on net8.0, net9.0, and net10.0.
